Fix handling of LODs and tile origin, add ArcGIS info endpoint

+++ handlers/tile.go
@@ -7,7 +7,6 @@ import (
 	"strings"
 )
 
-
 const (
 	earthRadius              = 6378137.0
 	earthCircumference       = math.Pi * earthRadius
@@ -63,7 +62,7 @@ func tileCoordFromString(z, x, y string) (tc tileCoord, ext string, err error) {
 	return
 }
 
-func calcScaleResolution(zoomLevel uint8, dpi uint8) (float64, float64) {
+func calcScaleResolution(zoomLevel int, dpi uint8) (float64, float64) {
 	var denom = 1 << zoomLevel
 	resolution := initialResolution / float64(denom)
 	scale := float64(dpi) * 39.37 * resolution // 39.37 in/m
